+++ tests/testthat/test_knn_sf.R
@@ -55,7 +55,7 @@ test_that("Testing the knn function with a simple case", {
   )
   exp_oid <- rbind(
     c(2,3),
-    c(3,1),
+    c(1,3),
     c(2,1)
   )
 
@@ -126,7 +126,7 @@ test_that("Testing the knn function with a simple case and a directed network", 
   )
   exp_oid <- rbind(
     c(3,2),
-    c(3,1),
+    c(1,3),
     c(2,1)
   )